Changed the folder structutre for Quiz module
