+++ src/app/api/teacher/course/module/[module_id]/quiz/[quiz_id]/route.ts
@@ -3,7 +3,7 @@ import { NextRequest } from "next/server";
 export async function GET(req:NextRequest)
 {
     try{
-
+        
     }
     catch(error)
     {
